+++ ms-pedidos/src/main/java/cl/pchardware/pedidos/repository/DetallePedidoRepository.java
@@ -1,12 +1,10 @@
 package cl.pchardware.pedidos.repository;
 
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import cl.pchardware.pedidos.model.DetallePedido;
 
 @Repository
-public interface DetallePedidoRepository extends JpaRepository<DetallePedido, Long> {
-    
+public interface DetallePedidoRepository extends JpaRepository<DetallePedido, Integer> {
 }
