+++ ms-pedidos/src/main/java/cl/pchardware/pedidos/repository/HistorialEstadoRepository.java
@@ -1,12 +1,10 @@
 package cl.pchardware.pedidos.repository;
 
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import cl.pchardware.pedidos.model.HistorialEstado;
 
 @Repository
-public interface HistorialEstadoRepository extends JpaRepository<HistorialEstado, Long> {
-    
+public interface HistorialEstadoRepository extends JpaRepository<HistorialEstado, Integer> {
 }
